evals/requirements.txt
anthropic>=0.25.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
rich>=13.0.0
pandas>=1.5.0
plotly>=5.0.0
streamlit>=1.28.0
